[webview] Pass Object instead of post-L webkit object over boundary.
When passing a ServiceWorkerWebSettings object across the support
library boundary - reference that object as an Object rather than as a
ServiceWorkerWebSettings.

The objective of the WebkitToCompatConverter is to convert webkit
objects into support library objects. But webkit objects introduced
post-L cannot be passed across the support library boundary since
whenever we create a Proxy for the interface
(WebkitToCompatConverterBoundaryInterface) passing the webkit object,
that Proxy needs access to the post-L webkit class (along with all other
classes referenced by the interface), which means we'll crash on phones
without a definition of the webkit class.

Bug: 831554
Change-Id: I12d731994a8d38dda0f69c92917c817ec753cfa4
Reviewed-on: https://chromium-review.googlesource.com/1006960
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Gustav Sennton <gsennton@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#549900}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4f96918a47768b2ae1e56cca65d63da8559af2f7