[supportlib] Support a :dev feature suffix.
:dev suffixed features will only be available if the device is
userdebug. This makes it possible to develop support library changes in
multiple steps, and then expose the feature with a final CL that removes
the suffix, without running the risk that applications will
inadvertently release code depending on the behaviour of an in-progress
implementation of a feature.

Bug: 908648
Change-Id: Iae98386992eb113356cf74be3eda125af10d867a
Reviewed-on: https://chromium-review.googlesource.com/c/1349327
Commit-Queue: Tobias Sargeant <tobiasjs@chromium.org>
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#611061}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c6e72efa6dba0e0ffff040451707dcbc71df3e89