[WebView Support Library] suppress lint warning
No change to logic.

This suppresses a lint warning (BanTargetApiAnnotation) which is
enforced in the AndroidX checkout (but not in Chromium). This is
required because this file is mirrored into AndroidX.

The intended resolution for this lint warning is to switch from
@TargetApi to @RequiresApi (which is not 1:1 compatible, but is
generally clearer). This is not currently possible in chromium, because
our lint doesn't seem to respect @RequiresApi like the AndroidX lint
config does.

R=torne@chromium.org, wnwen@chromium.org

Bug: http://b/120770118
Test: Manually copied into AndroidX checkout and ran "./gradlew
Test: lintMinDVD" (with http://aosp/848808)
Change-Id: Iad8574bf3fefa0ab0181b5618eb8a31f4b9ed182
Reviewed-on: https://chromium-review.googlesource.com/c/1374911
Reviewed-by: Peter Wen <wnwen@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#616395}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d652ab9a8b41cd708585690339ae3a6a234d6879