[AW] Change Prefetch thread requirement to be the UIThread
This CL changes the thread requirement for Prefetch methods to be the UI
to help prevent posting some of the pre-start logic to the UI thread.

This CL also changes how the cancel works since we pivoted away from
having a separate API to use CancellationSignal, so we pass along the
CancellationSignal instead of calling the cancel boundary interface
method.

This is a breaking change for the AndroidX API, hence changing the
feature name.

Bug: 355430425
Change-Id: I652da6bc6f086171c3e9fca949f917eef95d9776
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6105274
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Commit-Queue: Sayed Elabady <elabadysayed@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1398642}
NOKEYCHECK=True
GitOrigin-RevId: f3b8fe1dcc29059184deecf59dddb00c84891b8f