[AW] Change prefetch exception boundary interface
Instead of creating boundary interface for each value type of the
PrefetchExceptions, we can just use an enum and pass any necessary
information in the onFailure and let AndroidX create this exception based
on the data passed.

This should be the clean and optimized way of handling it.

Bug: 355430425
Change-Id: I186c3a443ecb24955442116ad49bb84a81b7f581
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6309156
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Commit-Queue: Sayed Elabady <elabadysayed@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1427142}
NOKEYCHECK=True
GitOrigin-RevId: 0434289fcb70ceaec09e711b2cb716640140bef5