webview: Add boundary interface for builder state configuration
Planning to experiment with a builder pattern to create a WebView.

Under the hood this will still initially construct a WebView with the
framework constructor (could re-think that in the future) but it would
give us the opportunity to configure the state that we know wouldn't
change and would be configured before WebView is used anywhere. From
AndroidX's perspective just gets a configured WebView back.

Adding a single configure boundary interface so that we can encode new
configurations into this, and not add a cost of adding new interfaces in
the future.

Build is always going to happen on the UI Thread so each configuration
will deal with its own validation.

Bug: b:409740828
Change-Id: I73fc218c75d1568eb8d33ee969c058ae736c4fa0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6448551
Reviewed-by: Ashley Newson <ashleynewson@chromium.org>
Reviewed-by: Peter Pakkenberg <pbirk@chromium.org>
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Commit-Queue: Rupert Wiser <bewise@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1454864}
NOKEYCHECK=True
GitOrigin-RevId: e666811c7198a7f0a4424516173c0f674bfa8c24