AW: Add support for multi-value Set-Cookie
The shouldInterceptRequest API allow the embedding application to
provide a Map<String, String> of response header values. The Set-Cookie
header does not support proper concatenation of values as a
comma-separated list.

To allow the client to set multiple cookie headers in the response, we
introduce a new unique header key, which will be allowed to contain
null-separated values in the value string.

We will provide a wrapper API as part of the androidx.webkit library to
allow clients to correctly set these values. The null-separation is only
meant to be an implementation detail.

Bug: b:414769380
Change-Id: Id6702d9860b9989944dba33b0b016aab36a36a94
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6494623
Reviewed-by: Adam Langley <agl@chromium.org>
Reviewed-by: Ashley Newson <ashleynewson@chromium.org>
Reviewed-by: Victor Tan <victortan@chromium.org>
Commit-Queue: Peter Pakkenberg <pbirk@chromium.org>
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1457583}
NOKEYCHECK=True
GitOrigin-RevId: f44aff5bdec42a7abb8727c6c03d5766295ebebc