AW: update comment about boundary_interfaces lint SDK
We would like the lint SDK to be hardcoded to exactly match the AndroidX
project. We don't want to use min_supported_sdk_version because this is
a value defined on the chromium side and it might change without
correlation to the AndroidX project. This updates the comment text to
provide more input about how this should work.

No change to logic. This is just for the sake of future-proofing.

Bug: 436287825
Test: N/A
Change-Id: Ia60290022fc73e495ab9be1b9ece38c60c7a91e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6825362
Reviewed-by: Benjamin Joyce (Ben) <bjoyce@chromium.org>
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Auto-Submit: Nate Fischer <ntfschr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1498344}
NOKEYCHECK=True
GitOrigin-RevId: 25066239641e92f1473f975dd32bf346e583f05d