[Reland] Change the API shape for BackForwardCacheSettings
This change alters the structure of BackForwardCacheSettings by grouping
the settings directly on the Settings object. This removes the need for
a builder and the corresponding setter method in the WebSettingsCompat
class.

This change was reverted due to a bug in the first version of the API
and this CL fixes it.

Bug: b:432395269
Change-Id: Ib3fd6e5c37997f10998e02b9635bd4b39bb76e28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7213581
Commit-Queue: Sayed Elabady <elabadysayed@chromium.org>
Reviewed-by: Ashley Newson <ashleynewson@chromium.org>
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1555528}
NOKEYCHECK=True
GitOrigin-RevId: b8c03f6e9492de9c404743ae072b40bd05f12a61